# api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/availability": {
      "get": {
        "operationId": "AvailabilityController_get",
        "parameters": [
          {
            "name": "from",
            "required": true,
            "in": "query",
            "schema": {
              "example": "2026-10-12",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "description": "Exclu",
            "schema": {
              "example": "2026-10-14",
              "type": "string"
            }
          },
          {
            "name": "room_type_id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Disponibilités et tarifs par type de chambre, nuit par nuit",
        "tags": [
          "Disponibilités"
        ]
      }
    },
    "/v1/reservations": {
      "get": {
        "operationId": "ReservationsController_list",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Jeton `nextCursor` de la page précédente",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fields",
            "required": false,
            "in": "query",
            "description": "Champs à renvoyer, séparés par des virgules (ex. id,status,guest)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Un ou plusieurs statuts séparés par des virgules",
            "schema": {
              "example": "confirmed,checked_in",
              "type": "string"
            }
          },
          {
            "name": "arrival_from",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2026-10-01",
              "type": "string"
            }
          },
          {
            "name": "arrival_to",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2026-10-31",
              "type": "string"
            }
          },
          {
            "name": "guest_id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "reference",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Recherche : référence, nom, prénom ou email du client (2 caractères minimum)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "departure_from",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2026-10-01",
              "type": "string"
            }
          },
          {
            "name": "departure_to",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2026-10-31",
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "arrival_date | -arrival_date | created_at | -created_at",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Réservations"
        ]
      },
      "post": {
        "operationId": "ReservationsController_create",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Clé unique par opération, conservée 24 h",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReservationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Créer une réservation (contrôle de disponibilité atomique)",
        "tags": [
          "Réservations"
        ]
      }
    },
    "/v1/reservations/quote": {
      "get": {
        "operationId": "ReservationsController_quote",
        "parameters": [
          {
            "name": "room_type_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "arrival_date",
            "required": true,
            "in": "query",
            "schema": {
              "example": "2026-10-12",
              "type": "string"
            }
          },
          {
            "name": "departure_date",
            "required": true,
            "in": "query",
            "schema": {
              "example": "2026-10-14",
              "type": "string"
            }
          },
          {
            "name": "adults",
            "required": true,
            "in": "query",
            "schema": {
              "example": 2,
              "type": "number"
            }
          },
          {
            "name": "children",
            "required": false,
            "in": "query",
            "schema": {
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "board",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "room_only",
                "half_board",
                "full_board"
              ]
            }
          },
          {
            "name": "exclude_reservation_id",
            "required": false,
            "in": "query",
            "description": "Réservation modifiée, exclue du calcul de disponibilité",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Prix et disponibilité d’un séjour envisagé",
        "tags": [
          "Réservations"
        ]
      }
    },
    "/v1/reservations/{id}": {
      "get": {
        "operationId": "ReservationsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Réservations"
        ]
      },
      "patch": {
        "operationId": "ReservationsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReservationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Réservations"
        ]
      }
    },
    "/v1/reservations/{id}/history": {
      "get": {
        "operationId": "ReservationsController_history",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Historique des opérations sur la réservation",
        "tags": [
          "Réservations"
        ]
      }
    },
    "/v1/reservations/{id}/room": {
      "put": {
        "operationId": "ReservationsController_assignRoom",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRoomDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Pré-attribuer une chambre (ou retirer l’attribution avec roomId: null)",
        "tags": [
          "Réservations"
        ]
      }
    },
    "/v1/reservations/{id}/cancel": {
      "post": {
        "operationId": "ReservationsController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelReservationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Réservations"
        ]
      }
    },
    "/v1/reservations/{id}/check-in": {
      "post": {
        "operationId": "ReservationsController_checkIn",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckInDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Arrivée : attribue une chambre propre (automatique ou choisie)",
        "tags": [
          "Réservations"
        ]
      }
    },
    "/v1/reservations/{id}/check-out": {
      "post": {
        "operationId": "ReservationsController_checkOut",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckOutDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Départ : la chambre passe en nettoyage",
        "tags": [
          "Réservations"
        ]
      }
    },
    "/v1/planning": {
      "get": {
        "operationId": "PlanningController_get",
        "parameters": [
          {
            "name": "from",
            "required": true,
            "in": "query",
            "schema": {
              "example": "2026-10-01",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "description": "Exclu ; 31 jours maximum",
            "schema": {
              "example": "2026-10-15",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Planning des chambres : chambres et séjours sur la période (31 jours max.)",
        "tags": [
          "Planning"
        ]
      }
    },
...
  },
  "info": {
    "title": "API – Système de gestion hôtelière",
    "description": "Socle V1 : session, menu dynamique, chambres, clients, disponibilités, réservations, réception.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateGuestDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "Amina"
          },
          "lastName": {
            "type": "string",
            "example": "Benali"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "example": "+33612345678"
          },
          "nationality": {
            "type": "string",
            "example": "FR",
            "description": "Code pays ISO 3166-1 alpha-2"
          },
          "segment": {
            "type": "string",
            "enum": [
              "individual",
              "group",
              "company"
            ]
          },
          "preferences": {
            "type": "object",
            "example": {
              "pillow": "firm",
              "floor": "high"
            }
          }
        },
        "required": [
          "firstName",
          "lastName"
        ]
      },
...
      "CreateReservationDto": {
        "type": "object",
        "properties": {
          "guestId": {
            "type": "string",
            "description": "Client existant (ou fournir `guest`)"
          },
          "guest": {
            "description": "Nouveau client créé avec la réservation",
            "allOf": [
              {
                "$ref": "#/components/schemas/CreateGuestDto"
              }
            ]
          },
          "roomTypeId": {
            "type": "string"
          },
          "arrivalDate": {
            "type": "string",
            "example": "2026-10-12"
          },
          "departureDate": {
            "type": "string",
            "example": "2026-10-14",
            "description": "Date de départ (nuit non incluse)"
          },
          "adults": {
            "type": "number",
            "example": 2
          },
          "children": {
            "type": "number",
            "example": 0
          },
          "board": {
            "type": "string",
            "enum": [
              "room_only",
              "half_board",
              "full_board"
            ]
          },
          "channel": {
            "type": "string",
            "enum": [
              "direct",
              "phone",
              "web",
              "agency",
              "ota"
            ]
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "roomTypeId",
          "arrivalDate",
          "departureDate",
          "adults",
          "board",
          "channel"
        ]
      },
      "UpdateReservationDto": {
        "type": "object",
        "properties": {
          "roomTypeId": {
            "type": "string"
          },
          "arrivalDate": {
            "type": "string"
          },
          "departureDate": {
            "type": "string"
          },
          "adults": {
            "type": "number"
          },
          "children": {
            "type": "number"
          },
          "board": {
            "type": "string",
            "enum": [
              "room_only",
              "half_board",
              "full_board"
            ]
          },
          "notes": {
            "type": "string"
          }
        }
      },
      "AssignRoomDto": {
        "type": "object",
        "properties": {
          "roomId": {
            "type": "object",
            "nullable": true,
            "description": "Chambre à pré-attribuer, ou null pour retirer l’attribution"
          }
        },
        "required": [
          "roomId"
        ]
      },
      "CancelReservationDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "example": "Demande du client"
          }
        },
        "required": [
          "reason"
        ]
      },
      "CheckInDto": {
        "type": "object",
        "properties": {
          "roomId": {
            "type": "string",
            "description": "Chambre choisie ; sinon attribution automatique"
          }
        }
      },
      "CheckOutDto": {
        "type": "object",
        "properties": {
          "overrideBalanceReason": {
            "type": "string",
            "description": "Départ malgré un solde non nul (profils habilités) ; motif obligatoire"
          }
        }
      },
...
    }
  }
}